refactor: MapsFragment 제거 및 지도 지도 관련 파일 이동
