+++ android/Staccato_AN/app/src/main/java/com/on/staccato/presentation/main/model/MarkerUiModel.kt
@@ -1,4 +1,4 @@
-package com.on.staccato.presentation.maps.model
+package com.on.staccato.presentation.main.model
 
 data class MarkerUiModel(
     val staccatoId: Long,
