+++ android/Staccato_AN/app/src/main/java/com/on/staccato/presentation/main/viewmodel/MapsViewModelFactory.kt
@@ -1,4 +1,4 @@
-package com.on.staccato.presentation.maps
+package com.on.staccato.presentation.main.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
