+++ android/Staccato_AN/app/src/main/java/com/on/staccato/presentation/main/viewmodel/SharedViewModel.kt
@@ -1,4 +1,4 @@
-package com.on.staccato.presentation.main
+package com.on.staccato.presentation.main.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
