--- android/Staccato_AN/app/src/main/java/com/on/staccato/presentation/maps/MapsFragment.kt
@@ -1,221 +1,0 @@
-package com.on.staccato.presentation.maps
-
-import android.Manifest
-import android.content.Intent
-import android.content.pm.PackageManager
-import android.location.Location
-import android.net.Uri
-import android.os.Bundle
-import android.provider.Settings
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.core.content.ContextCompat
-import androidx.core.os.bundleOf
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.activityViewModels
-import androidx.fragment.app.viewModels
-import androidx.navigation.NavOptions
-import androidx.navigation.fragment.findNavController
-import com.google.android.gms.location.FusedLocationProviderClient
-import com.google.android.gms.location.LocationRequest
-import com.google.android.gms.location.LocationServices
-import com.google.android.gms.maps.CameraUpdateFactory
-import com.google.android.gms.maps.GoogleMap
-import com.google.android.gms.maps.OnMapReadyCallback
-import com.google.android.gms.maps.SupportMapFragment
-import com.google.android.gms.maps.model.LatLng
-import com.google.android.gms.maps.model.Marker
-import com.google.android.gms.maps.model.MarkerOptions
-import com.google.android.material.bottomsheet.BottomSheetBehavior
-import com.google.android.material.snackbar.Snackbar
-import com.on.staccato.R
-import com.on.staccato.data.StaccatoClient.momentApiService
-import com.on.staccato.data.moment.MomentDefaultRepository
-import com.on.staccato.data.moment.MomentRemoteDataSource
-import com.on.staccato.domain.model.MomentLocation
-import com.on.staccato.presentation.main.SharedViewModel
-import com.on.staccato.presentation.maps.model.MarkerUiModel
-import com.on.staccato.presentation.moment.MomentFragment.Companion.MOMENT_ID_KEY
-
-class MapsFragment : Fragment() {
-    private val viewModel: MapsViewModel by viewModels {
-        MapsViewModelFactory(
-            MomentDefaultRepository(
-                MomentRemoteDataSource(momentApiService),
-            ),
-        )
-    }
-    private val sharedViewModel: SharedViewModel by activityViewModels<SharedViewModel>()
-    private lateinit var fusedLocationProviderClient: FusedLocationProviderClient
-
-    private val mapReadyCallback =
-        OnMapReadyCallback { googleMap ->
-            checkLocationPermissions(googleMap)
-            observeMomentLocations(googleMap)
-            onMarkerClicked(googleMap)
-        }
-
-    private val locationPermissions: Array<String> =
-        arrayOf(
-            Manifest.permission.ACCESS_FINE_LOCATION,
-            Manifest.permission.ACCESS_COARSE_LOCATION,
-        )
-
-    private val requestPermission = initRequestPermissionsLauncher()
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?,
-    ): View? {
-        fusedLocationProviderClient = LocationServices.getFusedLocationProviderClient(requireActivity())
-        return inflater.inflate(R.layout.fragment_maps, container, false)
-    }
-
-    override fun onViewCreated(
-        view: View,
-        savedInstanceState: Bundle?,
-    ) {
-        super.onViewCreated(view, savedInstanceState)
-        val mapFragment = childFragmentManager.findFragmentById(R.id.map) as SupportMapFragment?
-        mapFragment?.getMapAsync(mapReadyCallback)
-        observeStaccatoId()
-        observeDeletedStaccato()
-    }
-
-    override fun onResume() {
-        super.onResume()
-        viewModel.loadStaccatos()
-    }
-
-    private fun checkLocationPermissions(googleMap: GoogleMap) {
-        val isAccessFineLocationGranted =
-            ContextCompat.checkSelfPermission(
-                requireContext(),
-                Manifest.permission.ACCESS_FINE_LOCATION,
-            ) == PackageManager.PERMISSION_GRANTED
-
-        val isAccessCoarseLocationGranted =
-            ContextCompat.checkSelfPermission(
-                requireContext(),
-                Manifest.permission.ACCESS_COARSE_LOCATION,
-            ) == PackageManager.PERMISSION_GRANTED
-
-        if (isAccessFineLocationGranted || isAccessCoarseLocationGranted) {
-            googleMap.isMyLocationEnabled = true
-            val currentLocation =
-                fusedLocationProviderClient.getCurrentLocation(
-                    LocationRequest.PRIORITY_HIGH_ACCURACY,
-                    null,
-                )
-            currentLocation.addOnSuccessListener { location ->
-                moveCamera(googleMap, location)
-            }
-            return
-        } else {
-            requestPermission.launch(locationPermissions)
-        }
-    }
-
-    private fun initRequestPermissionsLauncher() =
-        registerForActivityResult(ActivityResultContracts.RequestMultiplePermissions()) { permissions ->
-            permissions.forEach { (_, isGranted) ->
-                if (isGranted) {
-                    makeSnackBar(getString(R.string.maps_location_permission_granted_message)).show()
-                } else {
-                    showPermissionRequestSnackBar()
-                }
-            }
-        }
-
-    private fun showPermissionRequestSnackBar() {
-        val snackBar = makeSnackBar(getString(R.string.maps_location_permission_required_message))
-        snackBar.setAction()
-        snackBar.show()
-    }
-
-    private fun makeSnackBar(message: String) = Snackbar.make(requireView(), message, Snackbar.LENGTH_LONG)
-
-    private fun Snackbar.setAction() {
-        setAction(R.string.snack_bar_move_to_setting) {
-            val uri = Uri.fromParts(PACKAGE_SCHEME, context.packageName, null)
-            val intent = Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS).setData(uri)
-            startActivity(intent)
-        }
-    }
-
-    private fun observeMomentLocations(googleMap: GoogleMap) {
-        viewModel.momentLocations.observe(viewLifecycleOwner) { momentLocations ->
-            googleMap.clear()
-            addMarkers(momentLocations, googleMap)
-        }
-    }
-
-    private fun addMarkers(
-        momentLocations: List<MomentLocation>,
-        googleMap: GoogleMap,
-    ) {
-        val markers: MutableList<MarkerUiModel> = mutableListOf()
-        momentLocations.forEach { momentLocation ->
-            val latLng = LatLng(momentLocation.latitude, momentLocation.longitude)
-            val markerOptions: MarkerOptions = MarkerOptions().position(latLng)
-            val marker: Marker = googleMap.addMarker(markerOptions) ?: return
-            val markerId: String = marker.id
-            markers.add(MarkerUiModel(momentLocation.momentId, markerId))
-        }
-        viewModel.setMarkers(markers)
-    }
-
-    private fun moveCamera(
-        googleMap: GoogleMap,
-        location: Location?,
-    ) {
-        if (location != null) {
-            val currentLocation = LatLng(location.latitude, location.longitude)
-            googleMap.moveCamera(CameraUpdateFactory.newLatLngZoom(currentLocation, 15f))
-        }
-    }
-
-    private fun onMarkerClicked(googleMap: GoogleMap) {
-        googleMap.setOnMarkerClickListener { marker ->
-            viewModel.findStaccatoId(marker.id)
-            false
-        }
-    }
-
-    private fun observeStaccatoId() {
-        viewModel.staccatoId.observe(viewLifecycleOwner) { staccatoId ->
-            navigateToStaccato(staccatoId)
-            parentFragmentManager.setFragmentResult(
-                BOTTOM_SHEET_STATE_REQUEST_KEY,
-                bundleOf(BOTTOM_SHEET_NEW_STATE to BottomSheetBehavior.STATE_EXPANDED),
-            )
-        }
-    }
-
-    private fun navigateToStaccato(staccatoId: Long?) {
-        val navOptions =
-            NavOptions.Builder()
-                .setPopUpTo(R.id.momentFragment, true)
-                .build()
-        val bundle =
-            bundleOf(MOMENT_ID_KEY to staccatoId)
-        findNavController().navigate(R.id.momentFragment, bundle, navOptions)
-    }
-
-    private fun observeDeletedStaccato() {
-        sharedViewModel.isStaccatosUpdated.observe(viewLifecycleOwner) { isDeleted ->
-            if (isDeleted) {
-                viewModel.loadStaccatos()
-            }
-        }
-    }
-
-    companion object {
-        const val PACKAGE_SCHEME = "package"
-        const val BOTTOM_SHEET_STATE_REQUEST_KEY = "requestKey"
-        const val BOTTOM_SHEET_NEW_STATE = "newState"
-    }
-}
